+++ NEWS.rst
@@ -59,6 +59,7 @@ Fixes
 
 Misc Improvements
 -----------------
+- `autogems`: Converted persistent data to json format
 - `catsplosion`: now a lua script instead of a plugin
 - `fix/diplomats`: replaces ``fixdiplomats``
 - `fix/merchants`: replaces ``fixmerchants``